requirements.txt
Django 
djangorestframework 
django-cors-headers 
djangorestframework-simplejwt
